requirements.txt
langchain
langchain-google-genai
pydantic
requests
pdfplumber
python-docx
python-dotenv
sentence-transformers
scikit-learn
fastapi
pdf2docx
pypandoc
jinja2
uvicorn